+++ CHANGELOG.md
@@ -3,4 +3,5 @@
 ## v1.0.34
 * Fixed proxy recipes incompatible with kjs recipe inject event.
 * Added support for rendering radius and global renderer.
+* Move trait dropping to the onDrop event.